scraping/wikipedia-en/requirements.txt
autopep8==2.0.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
Pillow==9.3.0
pycodestyle==2.10.0
requests==2.28.1
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.13
